fix multisig API error handling, fix header total balance
