+++ src/state/utils/walletStateUtils.ts
@@ -95,7 +95,7 @@ class WalletSessionToken{
       name: walletName,
       sub: "sessionLogin",
       iat: Math.floor(Date.now() / 1000),
-      exp: Math.floor(Date.now() / 1000) + (60 * 24 * 7), // 7 days
+      exp: Math.floor(Date.now() / 1000) + (60 * 60 * 3), // 3 hours
       salt: salt
     }
