+++ src/util/walletUtils.ts
@@ -543,9 +543,16 @@ export class WalletUtils {
         //WalletUtils.updateMultisigsDetails(wallet.accounts);
 
         for(let i = 0; i < wallet.accounts.length; ++i ){
-            let multisigInfo: MultisigInfo[] = await WalletUtils.getMultisigDetails(wallet.accounts[i].address);
 
-            wallet.accounts[i].multisigInfo = multisigInfo;
+            try {
+                let multisigInfo: MultisigInfo[] = await WalletUtils.getMultisigDetails(wallet.accounts[i].address);
+
+                wallet.accounts[i].multisigInfo = multisigInfo;
+            } catch (error) {
+                let multisigInfo: MultisigInfo[] = [];
+                multisigInfo.push(new MultisigInfo(wallet.accounts[i].publicKey, 0, [], [], 0, 0));
+                wallet.accounts[i].multisigInfo = multisigInfo;
+            }
         }
     }
 
@@ -561,9 +568,16 @@ export class WalletUtils {
         //WalletUtils.updateOtherAccountMultisigsDetails(wallet.others);
 
         for(let i = 0; i < wallet.others.length; ++i ){
-            let multisigInfo: MultisigInfo[] = await WalletUtils.getMultisigDetails(wallet.others[i].address);
 
-            wallet.others[i].multisigInfo = multisigInfo;
+            try {
+                let multisigInfo: MultisigInfo[] = await WalletUtils.getMultisigDetails(wallet.others[i].address);
+
+                wallet.others[i].multisigInfo = multisigInfo;
+            } catch (error) {
+                let multisigInfo: MultisigInfo[] = [];
+                multisigInfo.push(new MultisigInfo(wallet.others[i].publicKey, 0, [], [], 0, 0));
+                wallet.others[i].multisigInfo = multisigInfo;
+            }
         }
     }
 
@@ -644,11 +658,18 @@ export class WalletUtils {
 
         for(let i = 0; i < wallet.accounts.length; ++i){
 
+            if(wallet.accounts[i].multisigInfo === null){
+                continue;
+            }
+
             let publicKeys = wallet.accounts[i].getDirectChildMultisig();
 
             for(let i = 0; i < publicKeys.length; ++i){
 
                 if(wallet.others.find((other)=> other.publicKey === publicKeys[i])){
+                    continue;
+                }
+                else if(wallet.accounts.find((account)=> account.publicKey === publicKeys[i])){
                     continue;
                 }
 
@@ -893,7 +914,7 @@ export class WalletUtils {
             console.log(error);   
         }
 
-        walletState.wallets.savetoLocalStorage();
+        walletState.wallets.saveMyWalletOnlytoLocalStorage(wallet);
     }
 
     static updateAllAccountBalance(wallet: Wallet, assetId: string): void{
